+++ src/Contracts/ActiveCampaignResourceContract.php
@@ -2,9 +2,6 @@
 
 namespace Datomatic\ActiveCampaign\Contracts;
 
-/**
- * @property-read ActiveCampaignClientContract $client
- */
 interface ActiveCampaignResourceContract
 {
     public function client(): ActiveCampaignClientContract;