+++ src/Enums/ListStatus.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Datomatic\ActiveCampaign\Enums;
+
+enum ListStatus: int
+{
+    case Subscribed = 1;
+    case Unsubscribed = 2;
+}